ovn-controller: Support ovn-encap-ip-default option.
When there are multiple encap IPs configured for a chassis, there are
situations that any of the IP may be used, e.g. when encap-ip is not
configured for a VIF or when the output port of the pipeline is not
a VIF but a chassis-redirect port. In such cases, the encap IP used is
unpredictable.  This patch introduces the ovn-encap-ip-default option,
allowing the configuration of a default IP to be used to ensure
deterministic encap IP selection in such cases.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>