ci: Move DPDK build into container.
The DPDK was built as extra step in the CI, however it is useful to
have it inside the container prepared. This also helps with
reproduction of failures with DPDK by having the exact version inside
the container already.

Also bump the Ubuntu version for container builds to 24.04 to get
Podman 4. This is required to avoid the tar permissions error:

tar: dpdk-23.11/.ci: Cannot change mode to rwxrwxr-x: Operation not permitted

Signed-off-by: Ales Musil <amusil@redhat.com>
Acked-by: Eelco Chaudron <echaudro@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>