ovs: Bump submodule to branch-3.4.
Updating to OVS 3.4 to get enhanced sample action and other latest
features and fixes in general.

Unixctl library was changed to provide JSON-formatted output, so
clients has to be adjusted.  Clients also need to handle extra line
breaks at the end, since those are no longer prvided by the server.

We could define a common reply_to_string() function for both dbctl
and appctl, but these programs beahve differently in regerads to exit
codes and we'll need to do more refactoring to handle the difference.

Will move to v3.4.0 once it is available.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>