ovn-northd.at: Fix flaky VXLAN mode disabling test.
This tests constructs a very large operation that frequently takes more
than a second on busy CI systems.  As a result ovn-nbctl emits a warning
for a long poll interval and fails the test.

Fix that by splitting a single large command into a series of smaller
ones - 100 switches at a time.  This makes the test a bit slower, but
much more reliable.

Fixes: 7e99500e60bf ("northd: Add support for disabling vxlan mode.")
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>