binding: Cleanup related ports on port type change.
When a port type is changed from e.g. patch (which is a related_port)
to VIF which can be claimed by another chassis, then the port
should be removed from the related_ports set.

Test "ovn-controller local bindings" was already testing this
but not checking whether related_ports were correct.
This will be automatically tested when related_lports will be checked
at the end of the tests, in a following	patch.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>