github: Don't run scheduled workflows on forks.
It is annoying to have scheduled workflows running on your own fork,
especially if there are flaky tests or the workflow is broken and
permanently failing for an extended period of time generating an email
for every failure every week.

GitHub doesn't allow to selectively disable triggers, workflow can
only be disabled as a whole, so developers need to keep these failing
workflows running if they want to test their own changes before
submitting patches.

These scheduled workflows have little value on typical forks, i.e. the
ones that are not actual forks with custom code but forks for the
purpose of testing code before sending patches upstream.  Real forks
have code changes already, and they are free to modify the workflows
if they want the scheduled ones.

Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>