controller: Make sure the meter and group tables are initialized.
The following sequence of events could cause that the groups/meters
tables are not properly initialized:
1) Main loop runs and calls ovs_feature_support_run()
2) ovs_feature_support_run() returns true
3) ovs_feature_set_discovered() returns false, because not every
OpenFlow feature was discovered yet.
4) The inner loop for initialization doesn't run.
5) Main loop runs second time calling ovs_feature_support_run()
6) ovs_feature_support_run() returns false, because there wasn't
any feature state update.
7) ovs_feature_set_discovered() would return true, because every
OpenFlow feature was discovered with this run.
8) The inner if is never executed and the groups/meters
are not initialized.

Remove the check for full discovery to prevent that issue, at the
same time it wasn't needed anyway, the table would be updated only
if the maximum number of groups/meters would change during the update.

Fixes: 8c165db6f7d7 ("controller: fix group_table and meter_table allocation")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>