physical: Prevent wrong FDB to be learned with multichassis port.
If multichassis VIF is set with "unknown" address it might store
wrong MAC address into the FDB table. The ICMP Need frag is generated
by swapping the MAC src and dst of the original packet, however the
inport remains the same. As a consequence the match on the inport to
learn FDB will store source MAC address which is the original
destination address, that leads to redirection of traffic the VIF
which is wrong.

Swap the inport and outport for the ICMP error to prevent this issue
it also makes more sense as the ICMP is supposed to be generated by
entity along the way and not by the original VIF.

Note that those flows is still needed as userspace datapath is not
capable of PMTUD yet.

Reported-at: https://issues.redhat.com/browse/FDP-620
Signed-off-by: Ales Musil <amusil@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
(cherry picked from commit 5bf9cb91b6c4871a20a5c2df74c58451b492a5b9)
(cherry picked from commit 8b3e276e228407068e0a9e72ad3e02969da2de1d)