binding: Fix race condition when claiming vif.
When a vif is claimed, chassis is written in sb Port Binding.
If some vif update was received by the controller before the sb
notification of the chassis update (i.e. while pb->chassis = NULL),
the port was considered as TRACKED_REMOVED and ct_zone was flushed.
When, later, the pb->chassis update was received, ct_zone was flushed
again.
This issue seems to only have as side effect a few extra ct_zone flush while
the port is getting added.
The fix avoid the extra bump in the claim process and hence avoids
the unnecessary additional ct_flush.
This issue was causing some flaky failures of test
"Migration of CT zone from UUID to name" as unexpected ct_zone flus happened.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>