ovn-ic: Add support for route tag.
The OVN-IC feature can handle multiple points of redistribution when
we include more than one TS in the same Logical Router. So, it is a
good idea to use route tagging to prevent learing routes from logical
routers in the same user domain via Transit switch created for
peer-links with another user domain.

This means that there can be a logical separation by user domain
(e.g. VPCs), and a Transit Switch that peers between the different
VPCs, connecting the routers in the VPC1 with the routers in the VPC2,
for example. In this case, the route tag can ensure that the Logical
Routers of VPC1 does not learn routes from VPC1 itself via the peering
TS, avoiding installing additional openflow flows to handle
multipath/ecmp unnecessarily.

To use this route tag feature, new options were created to be configured
in the Logical Router Port connected to the TS. This configuration must
be performed in the LRP insted of the LR because the LRP is the path used
to advertise and learn routes connected to a specific TS (nexthop).

Signed-off-by: Roberto Bartzen Acosta <roberto.acosta@luizalabs.com>
Signed-off-by: Numan Siddique <numans@ovn.org>