northd: Fix NULL pointer deref within build_bfd_table.
The cb_bt pointer was pointing to different struct and might
have been NULL at this point. Make sure we are using the correct
pointer in the strcmp() and sbrec_bfd_set_chassis_name() call.

At the same time make sure we are actually running all tests that
do require the BFD beacon, as it was skipped in all CI stages. The
OpenBFDD project is not available as rpm, so we need to compile it
from source.

Fixes: 4885e337f929 ("rbac: Only allow relevant chassis to update BFD.")
Signed-off-by: Ales Musil <amusil@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>
(cherry picked from commit a55eb4a15f53c8169bd5f558ec2b1af0028e4d83)
Signed-off-by: Mark Michelson <mmichels@redhat.com>