tests: Fix typos found using check_uuid.
Next patch introduces check_uuid, checking if ovn-sbctl returns a uuid
(and not an error for instance).
This highlights a few tests issues, fixed in this patch.

Hostname has to be specified when creating chassis using sbctl to avoid
ovn-controller trying later to update hostname and cause RBAC permission
errors.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>