northd: Fix relying on naming coincidences.
We previous relied on the fact the the southbound Port_Binding gets the
same name as the Northbound port. However in all of the cases we use
this we already know the name the southbound port will have. Therefore we
just use this name, instead of getting it from northbound.

This is a prerequisite for later patches that use derived router ports.

Acked-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: Felix Huettner <felix.huettner@stackit.cloud>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>