Allow creation of a LRP without ipv4.
This makes it so that networks are an optional argument for lrp-add, and
in the case where no networks are provided it creates the logical router
port but doesn't add the routes that require ipv4.

This is useful as part of an effort to get OVN+BGP unnumbered working with
minimal configuration.

Signed-off-by: MJ Ponsonby <mj.ponsonby@canonical.com>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>