chassis: Check for ability to flush using CT label/mark.
This capability is not in all OVS versions, so we need to check for it
in order to use it.

A future commit will make use of this feature, so this lays the
groundwork for being able to confirm if chassis all support being able
to flush CT entries using label or mark bits.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Ales Musil <amusil@redhat.com>