treewide: Update OVS submodule to modernize SSL/TLS support.
This brings a whole pack of changes from OVS:

  * Dropped support for OpenSSL 1.1.0 and older.

    OpenSSL 1.1.0 reached EoL 5 years ago on 11 Sep 2019.  Vast
    majority of distributions moved to newer versions long time ago.

    OpenSSL 1.1.1 introduced a lot of new APIs and deprecated a lot of
    old ones.  It also introduced support for TLSv1.3 with a pack of
    APIs specific to that version.

    Requiring OpenSSL 1.1.1 or newer allowed OVS to introduce support
    for TLSv1.3 without a ton of conditional compiling as well as
    using more secure and easier adjustable default cipher list.

    Python community did an exceptional investigation on benefits of
    dropping support for OpenSSL 1.1.0 when they did the same in 2021:
      https://peps.python.org/pep-0644/

  * TLSv1 and TLSv1.1 are deprecated.  They will be removed from
    OVS 3.6, so marking as deprecated in OVN as well.

    TLSv1 and TLSv1.1 are officially deprecated by RFC 8996 since
    March of 2021:  https://datatracker.ietf.org/doc/rfc8996/

    Both protocols should not generally be used (RFC says MUST NOT)
    and are being actively removed from support by major distributions
    and libraries.

    Before this change, OVS/OVN would use TLSv1 or later, if the
    protocols are not specified in the database or command line (this
    includes TLSv1.3 that is not supported explicitly).  After the
    change, this becomes TLSv1.2 or later.

  * Added explicit support for TLSv1.3 and TLSv1.3+ ciphersuites.

  * Added support for ranges in ssl-protocols.

  * Defaults for ssl-ciphers with TLSv1.2 and earlier changed to
    'DEFAULT:@SECLEVEL=2' to get at least 112 bits of security
    strength and allow distributions to more easily control default
    set of ciphers to use by re-configuring OpenSSL.

There is one problem with setting database column ssl_ciphersuites
to be enabled for ovsdb-server in ovn-ctl script.  During upgrade,
if the old schema doesn't have this column, ovsdb-server will refuse
to start.  So, we can't turn this option on for clustered databases
by default.  We should be able to do so after next LTS release.
This should not be a big problem, since there should not be a high
demand for changing default ciphersuites (TLSv1.3 is relatively new
and all current ciphersuites are secure).  We could add an extra knob
to force enabling the option, but I'm not sure if that is necessary.
Users can ask for specific ciphsuites with a cmdline option instead
when necessary.
For standalone databases, schema upgrade is happening before starting
the ovsdb-server, so it's safe to turn the database option on.

Submodule is moved to point at OVS main, but the intention is to move
it to branch-3.5 as soon as it is available.

There are some line length warnings in ovn-ctl help strings, but those
need larger refactoring of all the surroundings if we want to get them
fit.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>